fix(mfa): logout for non-full sessions + pending TTL + atomic code consume + CLI audit + hardening
Security-review fixes for the Login MFA backend:

1. /api/logout and /api/logout-all now depend on get_current_session (any
   valid session kind) instead of get_current_user, so a user mid-MFA
   (mfa_pending/mfa_setup) can cancel. Audit actor_user_id uses sess.user_id.
2. Split the mfa_pending TTL from mfa_setup: new setting
   mfa_pending_ttl_minutes (default 5). create_session now takes a float
   ttl_hours; cookie max_age rounded to an integer.
3. Eliminate read-then-write races in /api/login/mfa: recovery codes are
   consumed with an atomic guarded UPDATE ... WHERE used_at IS NULL RETURNING
   (single-use even under concurrency); UserMfa is loaded FOR UPDATE so the
   TOTP anti-replay (last_used_step) check-then-set serializes. Same row-lock
   applied to /api/me/mfa/confirm for consistency.
4. CLI break-glass reset_user_mfa now writes an audit_log row
   (action="mfa.cli_reset", actor_user_id=None).
5. Fail-closed limiter on /api/login/mfa (check -> 503 + Retry-After:5) and
   defensive try/except around record_failure/reset.
6. Bound input lengths in schemas: CodeIn.code max_length=128,
   PasswordIn.password max_length=1024.

+++ backend/app/services/auth.py
@@ -37,7 +37,7 @@ class AuthService:
         self,
         user: User,
         *,
-        ttl_hours: int,
+        ttl_hours: float,
         kind: str = "full",
         ip: str | None = None,
         user_agent: str | None = None,
